+++ common/src/main/java/net/satisfy/vinery/core/registry/FlammableBlockRegistry.java
@@ -3,18 +3,19 @@ package net.satisfy.vinery.core.registry;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.FireBlock;
-import net.satisfy.vinery.platform.PlatformHelper;
+
+import static net.satisfy.vinery.core.registry.ObjectRegistry.*;
 
 public class FlammableBlockRegistry {
 
     public static void init() {
-        addFlammable(5, 20, ObjectRegistry.DARK_CHERRY_PLANKS.get(), ObjectRegistry.DARK_CHERRY_SLAB.get(), ObjectRegistry.DARK_CHERRY_STAIRS.get(), ObjectRegistry.DARK_CHERRY_FENCE.get(),
-                ObjectRegistry.DARK_CHERRY_FENCE_GATE.get());
+        addFlammable(5, 20, DARK_CHERRY_PLANKS.get(), DARK_CHERRY_SLAB.get(), DARK_CHERRY_STAIRS.get(), DARK_CHERRY_FENCE.get(),
+                DARK_CHERRY_FENCE_GATE.get());
 
-        addFlammable(5, 5, ObjectRegistry.STRIPPED_DARK_CHERRY_LOG.get(), ObjectRegistry.DARK_CHERRY_LOG.get(), ObjectRegistry.APPLE_LOG.get(),
-                ObjectRegistry.STRIPPED_DARK_CHERRY_WOOD.get(), ObjectRegistry.DARK_CHERRY_WOOD.get(), ObjectRegistry.APPLE_WOOD.get());
+        addFlammable(5, 5, STRIPPED_DARK_CHERRY_LOG.get(), DARK_CHERRY_LOG.get(), APPLE_LOG.get(),
+                STRIPPED_DARK_CHERRY_WOOD.get(), DARK_CHERRY_WOOD.get(), APPLE_WOOD.get());
 
-        addFlammable(30, 60, ObjectRegistry.DARK_CHERRY_LEAVES.get(), ObjectRegistry.GRAPEVINE_LEAVES.get());
+        addFlammable(30, 60, DARK_CHERRY_LEAVES.get(), GRAPEVINE_LEAVES.get(), APPLE_LEAVES.get());
     }
 
     public static void addFlammable(int burnOdd, int igniteOdd, Block... blocks) {